byte -> int
git-svn-id: http://www.castiello-tsai.org:8523/svn/L1JTW@251 96aee97a-5c11-ed42-8518-d7f33e96f5f9

+++ src/l1j/server/server/clientpackets/C_MoveChar.java
@@ -41,7 +41,7 @@ public class C_MoveChar extends ClientBasePacket {
 	// ■■■■■■■■■■■■■ 移動關連 ■■■■■■■■■■■
 	private static final byte HEADING_TABLE_X[] = { 0, 1, 1, 1, 0, -1, -1, -1 };// 4.26 Start
 	private static final byte HEADING_TABLE_Y[] = { -1, -1, 0, 1, 1, 1, 0, -1 };// 4.26 End
-	private static final byte CLIENT_LANGUAGE = Config.CLIENT_LANGUAGE; // 5.10
+	private static final int CLIENT_LANGUAGE = Config.CLIENT_LANGUAGE; // 5.10
 
 	// 移動
 	public C_MoveChar(byte[] decrypt, ClientThread client) throws Exception{
